# doc/swagger/swagger.yaml
basePath: /api/v1/
definitions:
  docs.CarRequest:
    properties:
      city_name:
        example: New York
        type: string
      hourly_rent_cost:
        example: 99.99
        type: number
      seats:
        example: 4
        type: integer
      status:
        example: Available
        type: string
      type:
        example: Luxury
        type: string
    type: object
  docs.CarResponse:
    properties:
      city_name:
        example: New York
        type: string
      hourly_rent_cost:
        example: 99.99
        type: number
      id:
        example: bdaf243e-b4d3-49d7-8be4-5ed1fb4dba0e
        type: string
      seats:
        example: 4
        type: integer
      status:
        example: Available
        type: string
      type:
        example: Luxury
        type: string
    type: object
  docs.ErrorBadRequest:
    properties:
      detail:
        example: bad request
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad Request
        type: string
    type: object
  docs.ErrorInternalServer:
    properties:
      detail:
        example: internal server error
        type: string
      status:
        example: 500
        type: integer
      title:
        example: Internal server error
        type: string
    type: object
  docs.ErrorInvCityName:
    properties:
      detail:
        example: city name is not valid
        type: string
      status:
        example: 400
        type: integer
      title:
        example: Bad request
        type: string
    type: object
...
  docs.ErrorNotFound:
    properties:
      detail:
        example: not found
        type: string
      status:
        example: 404
        type: integer
      title:
        example: Not Found
        type: string
    type: object
...
  docs.ListCarsResponse:
    properties:
      cars:
        items:
          $ref: '#/definitions/docs.CarResponse'
        type: array
    type: object
  docs.ListCitiesNameResponse:
    properties:
      cities_name:
        example:
        - Chicago
        - Los Angeles
        - New York
        items:
          type: string
        type: array
    type: object
...
host: localhost:5050
info:
  contact: {}
  description: This is an API to manage a car rent service
  title: Car Rent API
  version: "1.0"
paths:
  /cars:
    post:
      consumes:
      - application/json
      description: Register a new car with the provided information
      operationId: register-car
      parameters:
      - description: 'Car information (allowed types: Sedan, Luxury, Sports Car, Limousine;
          allowed statuses: Available, Unavailable)'
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/docs.CarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created car
          schema:
            $ref: '#/definitions/docs.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorInvCityName'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Register a new car
      tags:
      - Cars
  /cars/:
    get:
      description: |-
        Lists cars from a city in pages of 20 elements. from_car_id parameter
        is taken as the last seen car in a previous page.
      operationId: list-cars
      parameters:
      - description: City name
        in: query
        name: city
        required: true
        type: string
      - description: Last seen car ID
        format: uuid
        in: query
        name: from_car_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Obtained car
          schema:
            $ref: '#/definitions/docs.ListCarsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: List cars
      tags:
      - Cars
...
  /cars/{id}:
    delete:
      description: Delete a car by UUID
      operationId: delete-car
      parameters:
      - description: Car UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Delete a car
      tags:
      - Cars
    get:
      description: Get a car by UUID
      operationId: get-car
      parameters:
      - description: Car UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Obtained car
          schema:
            $ref: '#/definitions/docs.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Get a car
      tags:
      - Cars
    put:
      consumes:
      - application/json
      description: Update a car by UUID
      operationId: update-car
      parameters:
      - description: Car UUID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: 'Car information (allowed types: Sedan, Luxury, Sports Car, Limousine;
          allowed statuses: Available, Unavailable)'
        in: body
        name: car
        required: true
        schema:
          $ref: '#/definitions/docs.CarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated car
          schema:
            $ref: '#/definitions/docs.CarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/docs.ErrorBadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/docs.ErrorNotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: Update a car
      tags:
      - Cars
  /cities/names:
    get:
      description: Lists the names of all the currently supported cities
      operationId: list-cities
      produces:
      - application/json
      responses:
        "200":
          description: Cities name
          schema:
            $ref: '#/definitions/docs.ListCitiesNameResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/docs.ErrorInternalServer'
      summary: List cities
      tags:
      - Cities
...
swagger: "2.0"